redis - feat: moving all the types to a types file
